contract call trait - monomorphization fix
