+++ elrond-wasm/src/types/interaction/contract_call_trait.rs
@@ -1,11 +1,7 @@
 use elrond_codec::{multi_types::IgnoreValue, TopDecodeMulti, TopEncodeMulti};
 
 use crate::{
-    api::CallTypeApi,
-    contract_base::{ExitCodecErrorHandler, SendRawWrapper},
-    err_msg,
-    io::{ArgErrorHandler, ArgId, ManagedResultArgLoader},
-    types::{ManagedBuffer, ManagedVec},
+    api::CallTypeApi, contract_base::ExitCodecErrorHandler, err_msg, types::ManagedBuffer,
 };
 
 use super::{
@@ -50,32 +46,16 @@ where
     }
 
     fn async_call(self) -> AsyncCall<SA> {
-        let contract_call_full = self
-            .into_contract_call_full()
-            .convert_to_esdt_transfer_call();
-        AsyncCall {
-            to: contract_call_full.basic.to,
-            egld_payment: contract_call_full.egld_payment,
-            endpoint_name: contract_call_full.basic.endpoint_name,
-            arg_buffer: contract_call_full.basic.arg_buffer,
-            callback_call: None,
-        }
+        self.into_contract_call_full()
+            .convert_to_esdt_transfer_call()
+            .async_call()
     }
 
     #[cfg(feature = "promises")]
     fn async_call_promise(self) -> super::AsyncCallPromises<SA> {
-        let contract_call_full = self
-            .into_contract_call_full()
-            .convert_to_esdt_transfer_call();
-        super::AsyncCallPromises {
-            to: contract_call_full.basic.to,
-            egld_payment: contract_call_full.egld_payment,
-            endpoint_name: contract_call_full.basic.endpoint_name,
-            arg_buffer: contract_call_full.basic.arg_buffer,
-            explicit_gas_limit: contract_call_full.basic.explicit_gas_limit,
-            extra_gas_for_callback: 0,
-            callback_call: None,
-        }
+        self.into_contract_call_full()
+            .convert_to_esdt_transfer_call()
+            .async_call_promise()
     }
 
     /// Executes immediately, synchronously, and returns contract call result.
@@ -84,20 +64,9 @@ where
     where
         RequestedResult: TopDecodeMulti,
     {
-        let contract_call_full = self
-            .into_contract_call_full()
-            .convert_to_esdt_transfer_call();
-        let raw_result = SendRawWrapper::<SA>::new().execute_on_dest_context_raw(
-            contract_call_full.resolve_gas_limit(),
-            &contract_call_full.basic.to,
-            &contract_call_full.egld_payment,
-            &contract_call_full.basic.endpoint_name,
-            &contract_call_full.basic.arg_buffer,
-        );
-
-        SendRawWrapper::<SA>::new().clean_return_data();
-
-        decode_result(raw_result)
+        self.into_contract_call_full()
+            .convert_to_esdt_transfer_call()
+            .execute_on_dest_context()
     }
 
     /// Executes immediately, synchronously.
@@ -117,39 +86,18 @@ where
     where
         RequestedResult: TopDecodeMulti,
     {
-        let contract_call_full = self
-            .into_contract_call_full()
-            .convert_to_esdt_transfer_call();
-        let raw_result = SendRawWrapper::<SA>::new().execute_on_dest_context_readonly_raw(
-            contract_call_full.resolve_gas_limit(),
-            &contract_call_full.basic.to,
-            &contract_call_full.basic.endpoint_name,
-            &contract_call_full.basic.arg_buffer,
-        );
-
-        SendRawWrapper::<SA>::new().clean_return_data();
-
-        decode_result(raw_result)
+        self.into_contract_call_full()
+            .convert_to_esdt_transfer_call()
+            .execute_on_dest_context_readonly()
     }
 
     fn execute_on_same_context<RequestedResult>(self) -> RequestedResult
     where
         RequestedResult: TopDecodeMulti,
     {
-        let contract_call_full = self
-            .into_contract_call_full()
-            .convert_to_esdt_transfer_call();
-        let raw_result = SendRawWrapper::<SA>::new().execute_on_same_context_raw(
-            contract_call_full.resolve_gas_limit(),
-            &contract_call_full.basic.to,
-            &contract_call_full.egld_payment,
-            &contract_call_full.basic.endpoint_name,
-            &contract_call_full.basic.arg_buffer,
-        );
-
-        SendRawWrapper::<SA>::new().clean_return_data();
-
-        decode_result(raw_result)
+        self.into_contract_call_full()
+            .convert_to_esdt_transfer_call()
+            .execute_on_same_context()
     }
 
     /// Immediately launches a transfer-execute call.
@@ -157,26 +105,6 @@ where
     /// This is similar to an async call, but there is no callback
     /// and there can be more than one such call per transaction.
     fn transfer_execute(self) {
-        let contract_call_full = self.into_contract_call_full();
-
-        match contract_call_full.payments.len() {
-            0 => contract_call_full.no_payment_transfer_execute(),
-            1 => contract_call_full.single_transfer_execute(),
-            _ => contract_call_full.multi_transfer_execute(),
-        }
+        self.into_contract_call_full().transfer_execute();
     }
-}
-
-fn decode_result<SA, RequestedResult>(
-    raw_result: ManagedVec<SA, ManagedBuffer<SA>>,
-) -> RequestedResult
-where
-    SA: CallTypeApi + 'static,
-    RequestedResult: TopDecodeMulti,
-{
-    let mut loader = ManagedResultArgLoader::new(raw_result);
-    let arg_id = ArgId::from(&b"sync result"[..]);
-    let h = ArgErrorHandler::<SA>::from(arg_id);
-    let Ok(result) = RequestedResult::multi_decode_or_handle_err(&mut loader, h);
-    result
 }
